rename sprocess to satprocess
