+++ bin/landsat8
@@ -2,7 +2,7 @@
 
 import os
 import argparse
-from sprocess.landsat8 import Landsat8Scene as Scene
+from satprocess.landsat8 import Landsat8Scene as Scene
 
 '''
     Basic command line parser for processing
